+++ src/com/seafile/seadroid2/ui/WidgetUtils.java
@@ -4,22 +4,90 @@ import android.content.ActivityNotFoundException;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
+import android.content.pm.ResolveInfo;
 import android.net.Uri;
 import android.support.v4.app.FragmentActivity;
+import android.text.ClipboardManager;
 import android.webkit.MimeTypeMap;
+import com.actionbarsherlock.app.SherlockFragmentActivity;
 import com.nostra13.universalimageloader.core.DisplayImageOptions;
 import com.seafile.seadroid2.R;
 import com.seafile.seadroid2.SeadroidApplication;
+import com.seafile.seadroid2.account.Account;
 import com.seafile.seadroid2.ui.activity.BrowserActivity;
 import com.seafile.seadroid2.ui.activity.MarkdownActivity;
+import com.seafile.seadroid2.ui.dialog.AppChoiceDialog;
+import com.seafile.seadroid2.ui.dialog.GetShareLinkDialog;
 import com.seafile.seadroid2.ui.dialog.OpenAsDialog;
+import com.seafile.seadroid2.ui.dialog.TaskDialog;
+import com.seafile.seadroid2.util.Utils;
 
 import java.io.File;
+import java.util.List;
 
 /**
  * Activity Utils
  */
 public class WidgetUtils {
+
+    public static void chooseShareApp(final SherlockFragmentActivity activity,
+                                      final String repoID,
+                                      final String path,
+                                      final boolean isdir,
+                                      final Account account) {
+        final Intent shareIntent = new Intent();
+        shareIntent.setAction(Intent.ACTION_SEND);
+        shareIntent.setType("text/plain");
+
+        // Get a list of apps
+        List<ResolveInfo> infos = Utils.getAppsByIntent(shareIntent);
+
+        String title = activity.getString(isdir ? R.string.share_dir_link : R.string.share_file_link);
+
+        AppChoiceDialog dialog = new AppChoiceDialog();
+        dialog.addCustomAction(0, activity.getResources().getDrawable(R.drawable.copy_link),
+                activity.getString(R.string.copy_link));
+        dialog.init(title, infos, new AppChoiceDialog.OnItemSelectedListener() {
+            @Override
+            public void onCustomActionSelected(AppChoiceDialog.CustomAction action) {
+                final GetShareLinkDialog gdialog = new GetShareLinkDialog();
+                gdialog.init(repoID, path, isdir, account);
+                gdialog.setTaskDialogLisenter(new TaskDialog.TaskDialogListener() {
+                    @Override
+                    @SuppressWarnings("deprecation")
+                    public void onTaskSuccess() {
+                        ClipboardManager clipboard = (ClipboardManager)
+                                activity.getSystemService(Context.CLIPBOARD_SERVICE);
+                        clipboard.setText(gdialog.getLink());
+                        // ClipData clip = ClipData.newPlainText("seafile shared link", gdialog.getLink());
+                        // clipboard.setPrimaryClip(clip);
+                        ToastUtils.show(activity, R.string.link_ready_to_be_pasted);
+                    }
+                });
+                gdialog.show(activity.getSupportFragmentManager(), "DialogFragment");
+            }
+
+            @Override
+            public void onAppSelected(ResolveInfo appInfo) {
+                String className = appInfo.activityInfo.name;
+                String packageName = appInfo.activityInfo.packageName;
+                shareIntent.setClassName(packageName, className);
+
+                final GetShareLinkDialog gdialog = new GetShareLinkDialog();
+                gdialog.init(repoID, path, isdir, account);
+                gdialog.setTaskDialogLisenter(new TaskDialog.TaskDialogListener() {
+                    @Override
+                    public void onTaskSuccess() {
+                        shareIntent.putExtra(Intent.EXTRA_TEXT, gdialog.getLink());
+                        activity.startActivity(shareIntent);
+                    }
+                });
+                gdialog.show(activity.getSupportFragmentManager(), "DialogFragment");
+            }
+
+        });
+        dialog.show(activity.getSupportFragmentManager(), BrowserActivity.CHOOSE_APP_DIALOG_FRAGMENT_TAG);
+    }
 
     public static void showFile(final FragmentActivity activity, File file) {
